legacy/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
chardet>=5.0.0
